# EPIC 5: Complete UI Testing Infrastructure with Puppeteer MCP