--- a/pytest.ini
+++ b/pytest.ini
@@ -11,6 +11,14 @@ markers =
     integration: Tests that require database, Redis, or external APIs
     e2e: End-to-end tests that test full system behavior
     slow: Tests that take more than 1 second to run
+    load_test: Load and performance testing scenarios
+    api_resilience: API failure and resilience testing  
+    concurrent_test: Concurrency and race condition testing
+    resource_test: Resource exhaustion and limits testing
+    light_load: Light load testing scenarios (fast execution)
+    medium_load: Medium load testing scenarios (moderate execution time)
+    stress_load: Stress load testing scenarios (slow execution, high resource usage)
+    fast: Tests that complete in under 10 seconds
     
 # Coverage settings
 [coverage:run]