--- a/pytest.ini
+++ b/pytest.ini
@@ -19,6 +19,9 @@ markers =
     medium_load: Medium load testing scenarios (moderate execution time)
     stress_load: Stress load testing scenarios (slow execution, high resource usage)
     fast: Tests that complete in under 10 seconds
+    ui: UI tests requiring browser automation
+    visual: Visual regression tests
+    slow_ui: UI tests that take >10 seconds
     
 # Coverage settings
 [coverage:run]