--- a/test_results/message_burst_light_load_performance.json
+++ b/test_results/message_burst_light_load_performance.json
@@ -1,0 +1,10 @@
+{
+  "successful_messages": 46,
+  "error_rate": 0.08,
+  "throughput": 4.6,
+  "avg_response_time": 0.056541396223980446,
+  "benchmark_met": true,
+  "timestamp": "2025-06-26T22:22:15.801586",
+  "environment": "local",
+  "test_name": "message_burst_light_load"
+}