--- a/tests/config/__init__.py
+++ b/tests/config/__init__.py
@@ -1,0 +1,2 @@
+# tests/config/__init__.py
+"""Configuration package for resilience and performance testing."""