--- a/tests/ui/__init__.py
+++ b/tests/ui/__init__.py
@@ -1,0 +1,7 @@
+"""
+UI Testing Package for NADIA Dashboard
+
+This package contains browser automation tests using Puppeteer MCP for testing
+the NADIA dashboard interface, including visual regression testing and 
+cross-browser compatibility validation.
+"""