--- a/tests/ui/page_objects/__init__.py
+++ b/tests/ui/page_objects/__init__.py
@@ -1,0 +1,6 @@
+"""
+Page Object Models for NADIA Dashboard UI Testing
+
+This module contains page object classes that encapsulate dashboard functionality
+for reliable browser automation testing using Puppeteer MCP.
+"""