… .github/branch_protection_config.json
… CLAUDE.md
… scratchpads/issue-30-phase2-plan.md
… scratchpads/issue-30.md
… scratchpads/issue-34.md
… scratchpads/issue-38.md
… scratchpads/issue-43.md
… tests/test_api_resilience.py
… tests/test_concurrent_processing.py
… tests/test_protocol_api_integration.py
… tests/test_recovery_agent.py
… tests/test_resource_exhaustion.py
… tests/ui/README.md
… tests/ui/config.py
… tests/ui/conftest.py
… tests/ui/mcp_adapter.py
… tests/ui/page_objects/base_page.py
… tests/ui/page_objects/dashboard_page.py
… tests/ui/page_objects/review_page.py
… tests/ui/retry_utils.py
… tests/ui/screenshots/current/dashboard_basic_load.png
… tests/ui/screenshots/current/direct_test_screenshot.png
… tests/ui/screenshots/current/test_basic_functionality.png
… tests/ui/screenshots/current/test_framework_validation.png
… tests/ui/screenshots/current/validation_test.png
… tests/ui/security.py
… tests/ui/test_basic_screenshot.py
… tests/ui/test_code_review_improvements.py
… tests/ui/test_dashboard_core.py
… tests/ui/test_dashboard_interface.py