@RobeHGC 2025-06-27
## ✅ Code Review Improvements Implemented

All critical and recommended changes have been successfully implemented!

### 🚨 Critical Issues Addressed

✅ **Environment Validation**: Added health check fixture  
✅ **Condition-Based Waiting**: Replaced asyncio.sleep with proper waits  
✅ **Visual Tolerance**: Configurable screenshot comparison (10% default)  
✅ **Test Isolation**: Comprehensive cleanup between tests

### 🔧 Key Improvements

- Smart element waiting with visibility checking
- Health check prevents failures when dashboard unavailable  
- Visual regression with configurable tolerance
- Comprehensive test cleanup and state reset
- Enhanced error handling throughout

The UI testing infrastructure is now production-ready with significantly improved reliability! 🚀